requirements.txt
alembic==1.13.3
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
numpy>=2.0.0
pandas
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.4
gunicorn==21.2.0
Flask-CORS==4.0.0
pinecone==5.3.1
pinecone-client==5.0.1
sentence-transformers==3.1.1